Week7/auth-demo/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
pydantic==2.5.0
